# Optimize `ListPostings.Next()` and `ListPostings.At()`