--- a/tsdb/index/postings.go
+++ b/tsdb/index/postings.go
@@ -747,7 +747,7 @@ func (rp *removedPostings) Err() error {
 // ListPostings implements the Postings interface over a plain list.
 type ListPostings struct {
 	list []storage.SeriesRef
-	cur  storage.SeriesRef
+	pos  int
 }
 
 func NewListPostings(list []storage.SeriesRef) Postings {
@@ -759,39 +759,34 @@ func newListPostings(list ...storage.SeriesRef) *ListPostings {
 }
 
 func (it *ListPostings) At() storage.SeriesRef {
-	return it.cur
+	return it.list[it.pos-1]
 }
 
 func (it *ListPostings) Next() bool {
-	if len(it.list) > 0 {
-		it.cur = it.list[0]
-		it.list = it.list[1:]
+	if it.pos < len(it.list) {
+		it.pos++
 		return true
 	}
-	it.cur = 0
 	return false
 }
 
 func (it *ListPostings) Seek(x storage.SeriesRef) bool {
-	// If the current value satisfies, then return.
-	if it.cur >= x {
-		return true
+	if it.pos == 0 {
+		it.pos++
 	}
-	if len(it.list) == 0 {
+	if it.pos > len(it.list) {
 		return false
+	}
+	// If the current value satisfies, then return.
+	if it.list[it.pos-1] >= x {
+		return true
 	}
 
 	// Do binary search between current position and end.
-	i := sort.Search(len(it.list), func(i int) bool {
-		return it.list[i] >= x
-	})
-	if i < len(it.list) {
-		it.cur = it.list[i]
-		it.list = it.list[i+1:]
-		return true
-	}
-	it.list = nil
-	return false
+	it.pos = sort.Search(len(it.list[it.pos-1:]), func(i int) bool {
+		return it.list[i+it.pos-1] >= x
+	}) + it.pos
+	return it.pos-1 < len(it.list)
 }
 
 func (it *ListPostings) Err() error {