@bboreham 2023-10-09
This must make `At()` slower, surely?  It's doing an addition and a bounds-check now.
Often we will call `Next()` and `At()` the same number of times, but for instance `Merge()` does a lot more `At()` calls.