@colega 2023-10-09
@bboreham let me run the benchmarks